feat(manager) : add indicator column for marks
There are three major mark to indicate state of marks.
I set this before with separating modified/focused/altered variable.

but it is annoying. so remove these and add new variable `indicator`
which includes all three indicated marks.

update_marks() add/delete mark for list which not belong to marks.
so we needs this update separately