fix(manager) : focus current buffer if alternate buffer is hidden
If I open buffer manager right after remove current buffer using :bd,
alternate buffer will be shown and it is set as current buffer.
new alternate buffer is set as previous current buffer which is hidden
now.

In this situation, buffer manager cannot find alternate buffer because
it is hidden.
In this situation, current buffer will be focused in buffer manager.