fix : change bjump direction for `stack` sort method
For UX, change bjump direction for `stack` sort method

For other sort method, buffer manager shows the buffer list in ascending order.
It means that subordinate item is next buffer.

In stack sorting method, the newest item which is the last inserted
is pushed on the top of buffer list. after then, if other buffer is
visited,  the buffer is on the top.
The older buffer is pushed down in the stack.

If you want to visit previous item which is visited before, you think
bprev() function to go to them. Typically, it selects the item at
the top of the buffer list.
As is commonly thought, stack sorting method borrows the method of keeping bprev()
intact to return to the previous buffer.