fix : remove converting white space to '\\ ' in path. (sep_unify())
white space is converted by '\\ ' in sep_unify().
It is implemented to deal path which has white space.

In order to implement the buffer manager to properly extract each field even
if there is a path with spaces later.

But the method that converting %s to '\\ ' makes add \ in the middle of
filename or filepath. so If there are white space in the middle of file
name, It is sliced by \ when vim.fn.fnamemodify(':t') to get filename.

So it is restored to original code