fix : operate original behavior in edit_mode perfectly.
1. Problem

Previously, restoring key is implemented by using `vim.cmd('normal' .. key)`.
It intends to do original behavior If the key is registered as go to
buffer shortcut.
It works for single key stroke. but It doesn't to double key stroke like 'dd',
if there are shortcut 'd'.

2. Solution.

use nvim-feedkeys().
It inserts the key input to buffer without immediately execution. so
continuous input will be recognized as one command.